Conflicts:
	servers/src/main/java/tachyon/web/UiFileInfo.java
	servers/src/main/java/tachyon/worker/block/BlockDataManager.java
	servers/src/main/java/tachyon/worker/block/meta/StorageTier.java
	servers/src/test/java/tachyon/worker/block/BlockMetadataManagerViewTest.java

+++ servers/src/main/java/tachyon/worker/block/BlockStore.java
@@ -53,6 +53,10 @@ interface BlockStore {
 
   /**
    * NOTE: temporary, will be removed after changing client side code.
+   *
+   * @param userId the ID of the user to lock this block
+   * @param blockId the ID of the block to lock
+   * @throws NotFoundException if blockId can not be found, for example, evicted already.
    */
   void unlockBlock(long userId, long blockId) throws NotFoundException;
 
@@ -298,5 +302,5 @@ interface BlockStore {
    *
    * @param inodes a set of inodes that are currently pinned.
    */
-  void updatePinnedInodes(Set<Integer> inodes);
+  void updatePinnedInodes(Set<Long> inodes);
 }
